Fixes, refactoring, optimisation
* Adjusting engine for ClickHouse candles tables
* Removing dummy candles as obsolete
* Optimizing block download with reduced checks
* Enforcing void blocks download before repair attempts for ticks proto
- removing duplicate

+++ src/data_block.php
@@ -343,9 +343,11 @@
             $this->min_avail = $this->rbound;
             $this->history_fwd = [];
             $this->history_bwd = [];
-            if ($clean_cache) 
+            if ($clean_cache) {
                 foreach ($this->cache_files as $file_name)
-                    unlink($file_name);
+                    if (file_exists($file_name))
+                        unlink($file_name);
+            }
             $this->cache_files = [];
         }
